% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.snotel.r
\name{download.snotel}
\alias{download.snotel}
\title{Dowloads sno-tel data based upon a subset of the
sno-tel info as provided by snotel.info()}
\usage{
download.snotel(sites = NULL, path = ".")
}
\arguments{
\item{sites:}{subset of the sites listed by snotel.info()}

\item{path:}{where to save downloaded files}
}
\description{
Dowloads sno-tel data based upon a subset of the
sno-tel info as provided by snotel.info()
}
\examples{

# would download all available snotel data
# df = snotel.download(sites = snotel.info())
}
\keyword{SNOTEL,}
\keyword{USDA,}
\keyword{locations,}
\keyword{scraping}
\keyword{sites,}
\keyword{web}

