% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_snotel.r
\name{download_snotel}
\alias{download_snotel}
\title{Dowloads sno-tel data based upon a subset of the
sno-tel info as provided by snotel_info()}
\usage{
download_snotel(site = NULL, path = ".")
}
\arguments{
\item{site}{subset of the sites listed by snotel.info()}

\item{path}{where to save downloaded files}
}
\description{
Dowloads sno-tel data based upon a subset of the
sno-tel info as provided by snotel_info()
}
\examples{

# would download all available snotel data
# df = snotel_download(site = snotel_info())
}
\keyword{SNOTEL,}
\keyword{USDA,}
\keyword{locations,}
\keyword{scraping}
\keyword{sites,}
\keyword{web}
