% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel.explorer.r
\name{snotel.explorer}
\alias{snotel.explorer}
\title{Start the SNOTEL shiny interface}
\usage{
snotel.explorer()
}
\description{
Start the SNOTEL shiny interface
}
\examples{
# snotel.explorer()
}
\keyword{GUI}

