% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel.info.r
\name{snotel.info}
\alias{snotel.info}
\title{Grabs the sno-tel site listing for further processing}
\usage{

  snotel.info(url = "http://wcc.sc.egov.usda.gov/nwcc/yearcount?network=sntl&counttype=listwithdiscontinued&state=",
  path = NULL)
}
\arguments{
\item{url:}{Location of the FLUXNET2015 site table
(hopefully will not change to often, default is ok for now)}

\item{path:}{location of the phantomjs binary (system specific)}
}
\description{
Grabs the sno-tel site listing for further processing
}
\examples{

# with defaults, outputting a data frame
# df = snotel.info()

# [requires the rvest package for post-processing]
# http://phantomjs.org/download.html
}
\keyword{USDA,}
\keyword{locations,}
\keyword{scraping}
\keyword{sites,}
\keyword{sno-tel,}
\keyword{web}

