% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_explorer.r
\name{snotel_explorer}
\alias{snotel_explorer}
\title{Start the SNOTEL shiny interface}
\usage{
snotel_explorer()
}
\description{
Start the SNOTEL shiny interface
}
\examples{
# snotel_explorer()
}
\keyword{GUI}
