% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_metric.r
\name{snotel_metric}
\alias{snotel_metric}
\title{Convert snotel data to metric from imperial units}
\usage{
snotel_metric(df)
}
\arguments{
\item{df}{snotel data frame (or file)}
}
\description{
Data is read from either a snotel data file (csv) or 
a snotel data frame already loaded into the R workspace.
}
\details{
By default the conversion is done upon download. This function
might serve some a purpose in processing of data grabbed straight
from the server rather than through the package.

This is an internal function only. Hence, no examples are given.
}
\keyword{SNOTEL,}
\keyword{conversion,}
\keyword{metric}
