% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_phenology.r
\name{snotel_phenology}
\alias{snotel_phenology}
\title{Calculates snow phenology from the snow water
equivalent data as measured by the SNOTEL network.
first snow melt, first continuous snow melt, first
snow accumulation and continous snow accumulation
are reported.}
\usage{
snotel_phenology(df)
}
\arguments{
\item{df}{a snotel data file or data frame}
}
\description{
Calculates snow phenology from the snow water
equivalent data as measured by the SNOTEL network.
first snow melt, first continuous snow melt, first
snow accumulation and continous snow accumulation
are reported.
}
\examples{

# stats = estimate_phenology(df)
}
\keyword{SNOTEL,}
\keyword{USDA,}
\keyword{phenology,}
\keyword{series,}
\keyword{snow}
\keyword{time}
