% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_info.r
\name{snotel_info}
\alias{snotel_info}
\title{Downloads a SNOTEL site listing for further processing}
\usage{
snotel_info(path)
}
\arguments{
\item{path}{path where to save the snotel information (site list)}
}
\description{
Downloads a SNOTEL site listing for further processing
}
\examples{

\dontrun{
# download the meta-data from the SNOTEL server
meta_data <- snotel_info()

# show a couple of lines
head(meta_data)
}
}
\keyword{SNOTEL,}
\keyword{USDA,}
\keyword{locations,}
\keyword{scraping}
\keyword{sites,}
\keyword{web}
