% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_info.r
\name{snotel_info}
\alias{snotel_info}
\title{Downloads a SNOTEL site listing for further processing}
\usage{
snotel_info(network = "sntl", path)
}
\arguments{
\item{network}{network list to query (default = sntl, for SNOTEL)}

\item{path}{path where to save the snotel information (site list)}
}
\description{
Downloads a SNOTEL site listing for further processing
}
\examples{

\dontrun{
# download the meta-data from the SNOTEL server
meta_data <- snotel_info()

# show a couple of lines
head(meta_data)
}
}
