% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snotel_metric.r
\name{snotel_metric}
\alias{snotel_metric}
\title{Convert snotel data to metric from imperial units}
\usage{
snotel_metric(df, path = NULL)
}
\arguments{
\item{df}{snotel data frame or file}

\item{path}{where to save downloaded files}
}
\description{
Convert snotel data to metric from imperial units
}
\examples{

# would download all available snotel data
# df = snotel_metric(df = "snotel_1277.csv")
}
\keyword{SNOTEL,}
\keyword{USDA,}
\keyword{conversion}
\keyword{locations,}
\keyword{scraping,}
\keyword{sites,}
\keyword{web}
